cloud_func_code/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-pubsub>=0.28.1